move users to modules
