+++ server/users/package.json
@@ -34,7 +34,8 @@
     "passport-local": "^1.0.0",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
-    "rxjs": "^6.5.4"
+    "rxjs": "^6.5.4",
+    "shallow-equal-object": "^1.1.1"
   },
   "devDependencies": {
     "@nestjs/cli": "^7.0.0",
