+++ server/users/src/core/domain/value-object.ts
@@ -1,0 +1,23 @@
+import { shallowEqual } from 'shallow-equal-object';
+
+interface ValueObjectProps {
+  [index: string]: any;
+}
+
+export abstract class ValueObject<T extends ValueObjectProps> {
+  public readonly props: T;
+
+  constructor(props: T) {
+    this.props = Object.freeze(props);
+  }
+
+  public equals(vo?: ValueObject<T>): boolean {
+    if (vo === null || vo === undefined) {
+      return false;
+    }
+    if (vo.props === undefined) {
+      return false;
+    }
+    return shallowEqual(this.props, vo.props);
+  }
+}
