+++ server/users/src/modules/users/users.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { UsersController } from './controllers/users.controller';
 import { UsersService } from './services/users.service';
 import { UserRepository } from './repository/user.repository';
-import { HashingService } from './../utils/hashing.service';
+import { HashingService } from '../../utils/hashing.service';
 
 @Module({
   controllers: [UsersController],
