+++ server/users/test/users/users.e2e-spec.ts
@@ -1,8 +1,8 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { INestApplication, HttpStatus } from '@nestjs/common';
 import * as request from 'supertest';
-import { UsersModule } from '../../src/users/users.module';
-import { UserDto } from '../../src/users/dtos/user.dto';
+import { UsersModule } from '../../src/modules/users/users.module';
+import { UserDto } from '../../src/modules/users/dtos/user.dto';
 import { FakeDatabase } from '../../src/database/database.fake';
 
 describe('UsersController (e2e)', () => {
